+++ Reddit/SceneDelegate.swift
@@ -18,7 +18,8 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         
         if let windowScene = scene as? UIWindowScene {
             let window = UIWindow(windowScene: windowScene)
-            window.rootViewController = PermissionsContainerViewController()
+            let nav = UINavigationController(rootViewController: SearchViewController())
+            window.rootViewController = nav
             self.window = window
             window.makeKeyAndVisible()
         }